+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix the creation of two `EventSystem` on `Console` initialization [[#75](https://github.com/DarkRewar/BaseTool/issues/75)]
 - Fix the bug that blocks users to use `IfAttribute` with property accessor condition [[#78](https://github.com/DarkRewar/BaseTool/issues/78)]
+- Add unit tests to avoid regression [[87](https://github.com/DarkRewar/BaseTool/issues/78)]
 
 ## 0.4.0